Fixes for lint and clean up.
Delete unwanted code.
Clean up
Pass golintci
Start new README

+++ pkg/core/bundle_unpack.go
@@ -34,7 +34,7 @@ func unpackBundleFileList(ctx context.Context, bundle *Bundle) error {
 	for i = 0; i < bundle.BundleDescriptor.BundleEntriesFileCount; i++ {
 		bundleEntriesBuffer, err := storage.ReadTee(ctx,
 			bundle.ArchiveStore, model.GetArchivePathToBundleFileList(bundle.RepoID, bundle.BundleID, i),
-			bundle.ConsumableStore, model.GetConsumablePathToBundleFileList(bundle.BundleID, uint64(i)))
+			bundle.ConsumableStore, model.GetConsumablePathToBundleFileList(bundle.BundleID, i))
 		if err != nil {
 			return err
 		}
